Fix MSVC related warnings with sdl2

+++ Source/msg.cpp
@@ -770,7 +770,7 @@ void DeltaImportData(_cmd_id cmd, uint32_t recvOffset, int pnum)
 
 #ifdef USE_PKWARE
 	if (sgRecvBuf[0] != std::byte { 0 }) {
-		deltaSize = PkwareDecompress(&sgRecvBuf[1], deltaSize, sizeof(sgRecvBuf) - 1);
+		deltaSize = PkwareDecompress(&sgRecvBuf[1], static_cast<uint32_t>(deltaSize), sizeof(sgRecvBuf) - 1);
 		if (deltaSize == 0) {
 			Log("PKWare decompression failure, dropping player {}", pnum);
 			SNetDropPlayer(pnum, LEAVE_DROP);
@@ -3303,7 +3303,7 @@ bool ValidateCmdSize(size_t requiredCmdSize, size_t maxCmdSize, size_t playerId)
 		return true;
 
 	Log("Suspiciously small packet size, dropping player {}", playerId);
-	SNetDropPlayer(playerId, LEAVE_DROP);
+	SNetDropPlayer(static_cast<uint8_t>(playerId), LEAVE_DROP);
 	return false;
 }
